+++ packages/core/admin/admin/src/pages/Webhooks/EditView/components/EventInput/utils/formatValue.js
