+++ packages/core/admin/admin/src/pages/Webhooks/EditView/components/WebhookForm/tests/index.test.js
@@ -3,10 +3,10 @@ import userEvent from '@testing-library/user-event';
 import React from 'react';
 import { Router } from 'react-router';
 import { createMemoryHistory } from 'history';
-import en from '../../../../translations/en.json';
-import Theme from '../../../Theme';
+import en from '../../../../../../translations/en.json';
+import Theme from '../../../../../../components/Theme';
+import LanguageProvider from '../../../../../../components/LanguageProvider';
 import WebhookForm from '../index';
-import LanguageProvider from '../../../LanguageProvider';
 
 const makeApp = component => {
   const history = createMemoryHistory();
